% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_escolar_2019_docentes_centro_oeste}
\alias{censo_escolar_2019_docentes_centro_oeste}
\title{Docentes da educacao basica do Centro Oeste do Brasil - Censo Escolar 2019}
\description{
O censo_escolar_2019_docentes_centro_oeste contem diversos atributos de docentes da educacao basica do Centro Oeste do Brasil no ano de 2019. Os dados sao oriundos do censo escolar, levantamento de dados realizado pelo Ministerio da Educacao do Brasil (MEC) juntos as escolas e disponibilizado pelo INEP -  Instituto Nacional de Estudos e Pesquisas Educacionais Anisio Teixeira.
}
\usage{
ideb::censo_escolar_2019_docentes_centro_oeste
}
\format{
[,1] ano \cr
[,2] ID_DOCENTE \cr
[,3] NU_MES \cr
[,4] NU_ANO \cr
[,5] NU_IDADE_REFERENCIA \cr
[,6] NU_IDADE \cr
[,7] TP_SEXO \cr
[,8] TP_COR_RACA \cr
[,9] TP_NACIONALIDADE \cr
[,10] CO_PAIS_ORIGEM \cr
[,11] CO_UF_NASC \cr
[,12] CO_MUNICIPIO_NASC \cr
[,13] CO_UF_END \cr
[,14] CO_MUNICIPIO_END \cr
[,15] TP_ZONA_RESIDENCIAL \cr
[,16] TP_LOCAL_RESID_DIFERENCIADA \cr
[,17] IN_NECESSIDADE_ESPECIAL \cr
[,18] IN_BAIXA_VISAO \cr
[,19] IN_CEGUEIRA \cr
[,20] IN_DEF_AUDITIVA \cr
[,21] IN_DEF_FISICA \cr
[,22] IN_DEF_INTELECTUAL \cr
[,23] IN_SURDEZ \cr
[,24] IN_SURDOCEGUEIRA \cr
[,25] IN_DEF_MULTIPLA \cr
[,26] IN_AUTISMO \cr
[,27] IN_SUPERDOTACAO \cr
[,28] TP_ESCOLARIDADE \cr
[,29] TP_ENSINO_MEDIO \cr
[,30] TP_SITUACAO_CURSO_1 \cr
[,31] CO_AREA_CURSO_1 \cr
[,32] CO_CURSO_1 \cr
[,33] IN_LICENCIATURA_1 \cr
[,34] NU_ANO_CONCLUSAO_1 \cr
[,35] TP_TIPO_IES_1 \cr
[,36] CO_IES_1 \cr
[,37] TP_SITUACAO_CURSO_2 \cr
[,38] CO_AREA_CURSO_2 \cr
[,39] CO_CURSO_2 \cr
[,40] IN_LICENCIATURA_2 \cr
[,41] NU_ANO_CONCLUSAO_2 \cr
[,42] TP_TIPO_IES_2 \cr
[,43] CO_IES_2 \cr
[,44] TP_SITUACAO_CURSO_3 \cr
[,45] CO_AREA_CURSO_3 \cr
[,46] CO_CURSO_3 \cr
[,47] IN_LICENCIATURA_3 \cr
[,48] NU_ANO_CONCLUSAO_3 \cr
[,49] TP_TIPO_IES_3 \cr
[,50] CO_IES_3 \cr
[,51] IN_COMPLEMENTACAO_PEDAGOGICA \cr
[,52] CO_AREA_COMPL_PEDAGOGICA_1 \cr
[,53] CO_AREA_COMPL_PEDAGOGICA_2 \cr
[,54] CO_AREA_COMPL_PEDAGOGICA_3 \cr
[,55] IN_ESPECIALIZACAO \cr
[,56] IN_MESTRADO \cr
[,57] IN_DOUTORADO \cr
[,58] IN_POS_NENHUM \cr
[,59] IN_ESPECIFICO_CRECHE \cr
[,60] IN_ESPECIFICO_PRE_ESCOLA \cr
[,61] IN_ESPECIFICO_ANOS_INICIAIS \cr
[,62] IN_ESPECIFICO_ANOS_FINAIS \cr
[,63] IN_ESPECIFICO_ENS_MEDIO \cr
[,64] IN_ESPECIFICO_EJA \cr
[,65] IN_ESPECIFICO_ED_ESPECIAL \cr
[,66] IN_ESPECIFICO_ED_INDIGENA \cr
[,67] IN_ESPECIFICO_CAMPO \cr
[,68] IN_ESPECIFICO_AMBIENTAL \cr
[,69] IN_ESPECIFICO_DIR_HUMANOS \cr
[,70] IN_ESPECIFICO_DIV_SEXUAL \cr
[,71] IN_ESPECIFICO_DIR_ADOLESC \cr
[,72] IN_ESPECIFICO_AFRO \cr
[,73] IN_ESPECIFICO_GESTAO \cr
[,74] IN_ESPECIFICO_OUTROS \cr
[,75] IN_ESPECIFICO_NENHUM \cr
[,76] ID_TURMA \cr
[,77] TP_TIPO_DOCENTE \cr
[,78] TP_TIPO_CONTRATACAO \cr
[,79] IN_DISC_LINGUA_PORTUGUESA \cr
[,80] IN_DISC_EDUCACAO_FISICA \cr
[,81] IN_DISC_ARTES \cr
[,82] IN_DISC_LINGUA_INGLES \cr
[,83] IN_DISC_LINGUA_ESPANHOL \cr
[,84] IN_DISC_LINGUA_FRANCES \cr
[,85] IN_DISC_LINGUA_OUTRA \cr
[,86] IN_DISC_LIBRAS \cr
[,87] IN_DISC_LINGUA_INDIGENA \cr
[,88] IN_DISC_PORT_SEGUNDA_LINGUA \cr
[,89] IN_DISC_MATEMATICA \cr
[,90] IN_DISC_CIENCIAS \cr
[,91] IN_DISC_FISICA \cr
[,92] IN_DISC_QUIMICA \cr
[,93] IN_DISC_BIOLOGIA \cr
[,94] IN_DISC_HISTORIA \cr
[,95] IN_DISC_GEOGRAFIA \cr
[,96] IN_DISC_SOCIOLOGIA \cr
[,97] IN_DISC_FILOSOFIA \cr
[,98] IN_DISC_ESTUDOS_SOCIAIS \cr
[,99] IN_DISC_EST_SOCIAIS_SOCIOLOGIA \cr
[,100] IN_DISC_INFORMATICA_COMPUTACAO \cr
[,101] IN_DISC_ENSINO_RELIGIOSO \cr
[,102] IN_DISC_PROFISSIONALIZANTE \cr
[,103] IN_DISC_ESTAGIO_SUPERVISIONADO \cr
[,104] IN_DISC_PEDAGOGICAS \cr
[,105] IN_DISC_OUTRAS \cr
[,106] TP_TIPO_ATENDIMENTO_TURMA \cr
[,107] TP_TIPO_LOCAL_TURMA \cr
[,108] TP_MEDIACAO_DIDATICO_PEDAGO \cr
[,109] TP_ETAPA_ENSINO \cr
[,110] CO_CURSO_EDUC_PROFISSIONAL \cr
[,111] IN_ESPECIAL_EXCLUSIVA \cr
[,112] IN_REGULAR \cr
[,113] IN_EJA \cr
[,114] IN_PROFISSIONALIZANTE \cr
[,115] id_escola \cr
[,116] cod_regiao \cr
[,117] cod_mesorregiao \cr
[,118] cod_microrregiao \cr
[,119] cod_UF \cr
[,120] cod_municipio \cr
[,121] cod_distrito \cr
[,122] TP_DEPENDENCIA \cr
[,123] TP_LOCALIZACAO \cr
[,124] TP_CATEGORIA_ESCOLA_PRIVADA \cr
[,125] IN_CONVENIADA_PP \cr
[,126] TP_CONVENIO_PODER_PUBLICO \cr
[,127] IN_MANT_ESCOLA_PRIVADA_EMP \cr
[,128] IN_MANT_ESCOLA_PRIVADA_ONG \cr
[,129] IN_MANT_ESCOLA_PRIVADA_OSCIP \cr
[,130] IN_MANT_ESCOLA_PRIV_ONG_OSCIP \cr
[,131] IN_MANT_ESCOLA_PRIVADA_SIND \cr
[,132] IN_MANT_ESCOLA_PRIVADA_SIST_S \cr
[,133] IN_MANT_ESCOLA_PRIVADA_S_FINS \cr
[,134] TP_REGULAMENTACAO \cr
[,135] TP_LOCALIZACAO_DIFERENCIADA \cr
[,136] IN_EDUCACAO_INDIGENA
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
censo_escolar_2019_docentes_centro_oeste
}
\keyword{datasets}

