% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_escolar_2019_escolas}
\alias{censo_escolar_2019_escolas}
\title{Escolas da educacao basica do Brasil - Censo Escolar 2019}
\description{
O censo_escolar_2019_escolas contem diversos atributos de escolas da educacao basica do Brasil no ano de 2019. Os dados sao oriundos do censo escolar, levantamento de dados realizado pelo Ministerio da Educacao do Brasil (MEC) juntos as escolas e disponibilizado pelo INEP -  Instituto Nacional de Estudos e Pesquisas Educacionais Anisio Teixeira.
}
\usage{
ideb::censo_escolar_2019_escolas
}
\format{
[,1] ano \cr
[,2] id_escola \cr
[,3] nome_escola \cr
[,4] CO_ORGAO_REGIONAL \cr
[,5] TP_SITUACAO_FUNCIONAMENTO \cr
[,6] DT_ANO_LETIVO_INICIO \cr
[,7] DT_ANO_LETIVO_TERMINO \cr
[,8] cod_regiao \cr
[,9] cod_mesorregiao \cr
[,10] cod_microrregiao \cr
[,11] cod_UF \cr
[,12] cod_municipio \cr
[,13] cod_distrito \cr
[,14] TP_DEPENDENCIA \cr
[,15] TP_LOCALIZACAO \cr
[,16] TP_LOCALIZACAO_DIFERENCIADA \cr
[,17] IN_VINCULO_SECRETARIA_EDUCACAO \cr
[,18] IN_VINCULO_SEGURANCA_PUBLICA \cr
[,19] IN_VINCULO_SECRETARIA_SAUDE \cr
[,20] IN_VINCULO_OUTRO_ORGAO \cr
[,21] TP_CATEGORIA_ESCOLA_PRIVADA \cr
[,22] IN_CONVENIADA_PP \cr
[,23] TP_CONVENIO_PODER_PUBLICO \cr
[,24] IN_MANT_ESCOLA_PRIVADA_EMP \cr
[,25] IN_MANT_ESCOLA_PRIVADA_ONG \cr
[,26] IN_MANT_ESCOLA_PRIVADA_OSCIP \cr
[,27] IN_MANT_ESCOLA_PRIV_ONG_OSCIP \cr
[,28] IN_MANT_ESCOLA_PRIVADA_SIND \cr
[,29] IN_MANT_ESCOLA_PRIVADA_SIST_S \cr
[,30] IN_MANT_ESCOLA_PRIVADA_S_FINS \cr
[,31] TP_REGULAMENTACAO \cr
[,32] TP_RESPONSAVEL_REGULAMENTACAO \cr
[,33] CO_ESCOLA_SEDE_VINCULADA \cr
[,34] CO_IES_OFERTANTE \cr
[,35] IN_LOCAL_FUNC_PREDIO_ESCOLAR \cr
[,36] TP_OCUPACAO_PREDIO_ESCOLAR \cr
[,37] IN_LOCAL_FUNC_SOCIOEDUCATIVO \cr
[,38] IN_LOCAL_FUNC_UNID_PRISIONAL \cr
[,39] IN_LOCAL_FUNC_PRISIONAL_SOCIO \cr
[,40] IN_LOCAL_FUNC_GALPAO \cr
[,41] TP_OCUPACAO_GALPAO \cr
[,42] IN_LOCAL_FUNC_SALAS_OUTRA_ESC \cr
[,43] IN_LOCAL_FUNC_OUTROS \cr
[,44] IN_PREDIO_COMPARTILHADO \cr
[,45] IN_AGUA_POTAVEL \cr
[,46] IN_AGUA_REDE_PUBLICA \cr
[,47] IN_AGUA_POCO_ARTESIANO \cr
[,48] IN_AGUA_CACIMBA \cr
[,49] IN_AGUA_FONTE_RIO \cr
[,50] IN_AGUA_INEXISTENTE \cr
[,51] IN_ENERGIA_REDE_PUBLICA \cr
[,52] IN_ENERGIA_GERADOR_FOSSIL \cr
[,53] IN_ENERGIA_RENOVAVEL \cr
[,54] IN_ENERGIA_INEXISTENTE \cr
[,55] IN_ESGOTO_REDE_PUBLICA \cr
[,56] IN_ESGOTO_FOSSA_SEPTICA \cr
[,57] IN_ESGOTO_FOSSA_COMUM \cr
[,58] IN_ESGOTO_FOSSA \cr
[,59] IN_ESGOTO_INEXISTENTE \cr
[,60] IN_LIXO_SERVICO_COLETA \cr
[,61] IN_LIXO_QUEIMA \cr
[,62] IN_LIXO_ENTERRA \cr
[,63] IN_LIXO_DESTINO_FINAL_PUBLICO \cr
[,64] IN_LIXO_DESCARTA_OUTRA_AREA \cr
[,65] IN_TRATAMENTO_LIXO_SEPARACAO \cr
[,66] IN_TRATAMENTO_LIXO_REUTILIZA \cr
[,67] IN_TRATAMENTO_LIXO_RECICLAGEM \cr
[,68] IN_TRATAMENTO_LIXO_INEXISTENTE \cr
[,69] IN_ALMOXARIFADO \cr
[,70] IN_AREA_VERDE \cr
[,71] IN_AUDITORIO \cr
[,72] IN_BANHEIRO \cr
[,73] IN_BANHEIRO_EI \cr
[,74] IN_BANHEIRO_PNE \cr
[,75] IN_BANHEIRO_FUNCIONARIOS \cr
[,76] IN_BANHEIRO_CHUVEIRO \cr
[,77] IN_BIBLIOTECA \cr
[,78] IN_BIBLIOTECA_SALA_LEITURA \cr
[,79] IN_COZINHA \cr
[,80] IN_DESPENSA \cr
[,81] IN_DORMITORIO_ALUNO \cr
[,82] IN_DORMITORIO_PROFESSOR \cr
[,83] IN_LABORATORIO_CIENCIAS \cr
[,84] IN_LABORATORIO_INFORMATICA \cr
[,85] IN_PATIO_COBERTO \cr
[,86] IN_PATIO_DESCOBERTO \cr
[,87] IN_PARQUE_INFANTIL \cr
[,88] IN_PISCINA \cr
[,89] IN_QUADRA_ESPORTES \cr
[,90] IN_QUADRA_ESPORTES_COBERTA \cr
[,91] IN_QUADRA_ESPORTES_DESCOBERTA \cr
[,92] IN_REFEITORIO \cr
[,93] IN_SALA_ATELIE_ARTES \cr
[,94] IN_SALA_MUSICA_CORAL \cr
[,95] IN_SALA_ESTUDIO_DANCA \cr
[,96] IN_SALA_MULTIUSO \cr
[,97] IN_SALA_DIRETORIA \cr
[,98] IN_SALA_LEITURA \cr
[,99] IN_SALA_PROFESSOR \cr
[,100] IN_SALA_REPOUSO_ALUNO \cr
[,101] IN_SECRETARIA \cr
[,102] IN_SALA_ATENDIMENTO_ESPECIAL \cr
[,103] IN_TERREIRAO \cr
[,104] IN_VIVEIRO \cr
[,105] IN_DEPENDENCIAS_OUTRAS \cr
[,106] IN_ACESSIBILIDADE_CORRIMAO \cr
[,107] IN_ACESSIBILIDADE_ELEVADOR \cr
[,108] IN_ACESSIBILIDADE_PISOS_TATEIS \cr
[,109] IN_ACESSIBILIDADE_VAO_LIVRE \cr
[,110] IN_ACESSIBILIDADE_RAMPAS \cr
[,111] IN_ACESSIBILIDADE_SINAL_SONORO \cr
[,112] IN_ACESSIBILIDADE_SINAL_TATIL \cr
[,113] IN_ACESSIBILIDADE_SINAL_VISUAL \cr
[,114] IN_ACESSIBILIDADE_INEXISTENTE \cr
[,115] QT_SALAS_UTILIZADAS_DENTRO \cr
[,116] QT_SALAS_UTILIZADAS_FORA \cr
[,117] QT_SALAS_UTILIZADAS \cr
[,118] QT_SALAS_UTILIZA_CLIMATIZADAS \cr
[,119] QT_SALAS_UTILIZADAS_ACESSIVEIS \cr
[,120] IN_EQUIP_PARABOLICA \cr
[,121] IN_COMPUTADOR \cr
[,122] IN_EQUIP_COPIADORA \cr
[,123] IN_EQUIP_IMPRESSORA \cr
[,124] IN_EQUIP_IMPRESSORA_MULT \cr
[,125] IN_EQUIP_SCANNER \cr
[,126] IN_EQUIP_DVD \cr
[,127] QT_EQUIP_DVD \cr
[,128] IN_EQUIP_SOM \cr
[,129] QT_EQUIP_SOM \cr
[,130] IN_EQUIP_TV \cr
[,131] QT_EQUIP_TV \cr
[,132] IN_EQUIP_LOUSA_DIGITAL \cr
[,133] QT_EQUIP_LOUSA_DIGITAL \cr
[,134] IN_EQUIP_MULTIMIDIA \cr
[,135] QT_EQUIP_MULTIMIDIA \cr
[,136] IN_DESKTOP_ALUNO \cr
[,137] QT_DESKTOP_ALUNO \cr
[,138] IN_COMP_PORTATIL_ALUNO \cr
[,139] QT_COMP_PORTATIL_ALUNO \cr
[,140] IN_TABLET_ALUNO \cr
[,141] QT_TABLET_ALUNO \cr
[,142] IN_INTERNET \cr
[,143] IN_INTERNET_ALUNOS \cr
[,144] IN_INTERNET_ADMINISTRATIVO \cr
[,145] IN_INTERNET_APRENDIZAGEM \cr
[,146] IN_INTERNET_COMUNIDADE \cr
[,147] IN_ACESSO_INTERNET_COMPUTADOR \cr
[,148] IN_ACES_INTERNET_DISP_PESSOAIS \cr
[,149] TP_REDE_LOCAL \cr
[,150] IN_BANDA_LARGA \cr
[,151] QT_PROF_ADMINISTRATIVOS \cr
[,152] QT_PROF_SERVICOS_GERAIS \cr
[,153] QT_PROF_BIBLIOTECARIO \cr
[,154] QT_PROF_SAUDE \cr
[,155] QT_PROF_COORDENADOR \cr
[,156] QT_PROF_FONAUDIOLOGO \cr
[,157] QT_PROF_NUTRICIONISTA \cr
[,158] QT_PROF_PSICOLOGO \cr
[,159] QT_PROF_ALIMENTACAO \cr
[,160] QT_PROF_PEDAGOGIA \cr
[,161] QT_PROF_SECRETARIO \cr
[,162] QT_PROF_SEGURANCA \cr
[,163] QT_PROF_MONITORES \cr
[,164] IN_ALIMENTACAO \cr
[,165] IN_SERIE_ANO \cr
[,166] IN_PERIODOS_SEMESTRAIS \cr
[,167] IN_FUNDAMENTAL_CICLOS \cr
[,168] IN_GRUPOS_NAO_SERIADOS \cr
[,169] IN_MODULOS \cr
[,170] IN_FORMACAO_ALTERNANCIA \cr
[,171] IN_MATERIAL_PED_MULTIMIDIA \cr
[,172] IN_MATERIAL_PED_INFANTIL \cr
[,173] IN_MATERIAL_PED_CIENTIFICO \cr
[,174] IN_MATERIAL_PED_DIFUSAO \cr
[,175] IN_MATERIAL_PED_MUSICAL \cr
[,176] IN_MATERIAL_PED_JOGOS \cr
[,177] IN_MATERIAL_PED_ARTISTICAS \cr
[,178] IN_MATERIAL_PED_DESPORTIVA \cr
[,179] IN_MATERIAL_PED_INDIGENA \cr
[,180] IN_MATERIAL_PED_ETNICO \cr
[,181] IN_MATERIAL_PED_CAMPO \cr
[,182] IN_EDUCACAO_INDIGENA \cr
[,183] TP_INDIGENA_LINGUA \cr
[,184] CO_LINGUA_INDIGENA_1 \cr
[,185] CO_LINGUA_INDIGENA_2 \cr
[,186] CO_LINGUA_INDIGENA_3 \cr
[,187] IN_EXAME_SELECAO \cr
[,188] IN_RESERVA_PPI \cr
[,189] IN_RESERVA_RENDA \cr
[,190] IN_RESERVA_PUBLICA \cr
[,191] IN_RESERVA_PCD \cr
[,192] IN_RESERVA_OUTROS \cr
[,193] IN_RESERVA_NENHUMA \cr
[,194] IN_REDES_SOCIAIS \cr
[,195] IN_ESPACO_ATIVIDADE \cr
[,196] IN_ESPACO_EQUIPAMENTO \cr
[,197] IN_ORGAO_ASS_PAIS \cr
[,198] IN_ORGAO_ASS_PAIS_MESTRES \cr
[,199] IN_ORGAO_CONSELHO_ESCOLAR \cr
[,200] IN_ORGAO_GREMIO_ESTUDANTIL \cr
[,201] IN_ORGAO_OUTROS \cr
[,202] IN_ORGAO_NENHUM \cr
[,203] TP_PROPOSTA_PEDAGOGICA \cr
[,204] TP_AEE \cr
[,205] TP_ATIVIDADE_COMPLEMENTAR \cr
[,206] IN_MEDIACAO_PRESENCIAL \cr
[,207] IN_MEDIACAO_SEMIPRESENCIAL \cr
[,208] IN_MEDIACAO_EAD \cr
[,209] IN_ESPECIAL_EXCLUSIVA \cr
[,210] IN_REGULAR \cr
[,211] IN_EJA \cr
[,212] IN_PROFISSIONALIZANTE \cr
[,213] IN_COMUM_CRECHE \cr
[,214] IN_COMUM_PRE \cr
[,215] IN_COMUM_FUND_AI \cr
[,216] IN_COMUM_FUND_AF \cr
[,217] IN_COMUM_MEDIO_MEDIO \cr
[,218] IN_COMUM_MEDIO_INTEGRADO \cr
[,219] IN_COMUM_MEDIO_NORMAL \cr
[,220] IN_ESP_EXCLUSIVA_CRECHE \cr
[,221] IN_ESP_EXCLUSIVA_PRE \cr
[,222] IN_ESP_EXCLUSIVA_FUND_AI \cr
[,223] IN_ESP_EXCLUSIVA_FUND_AF \cr
[,224] IN_ESP_EXCLUSIVA_MEDIO_MEDIO \cr
[,225] IN_ESP_EXCLUSIVA_MEDIO_INTEGR \cr
[,226] IN_ESP_EXCLUSIVA_MEDIO_NORMAL \cr
[,227] IN_COMUM_EJA_FUND \cr
[,228] IN_COMUM_EJA_MEDIO \cr
[,229] IN_COMUM_EJA_PROF \cr
[,230] IN_ESP_EXCLUSIVA_EJA_FUND \cr
[,231] IN_ESP_EXCLUSIVA_EJA_MEDIO \cr
[,232] IN_ESP_EXCLUSIVA_EJA_PROF \cr
[,233] IN_COMUM_PROF \cr
[,234] IN_ESP_EXCLUSIVA_PROF
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
censo_escolar_2019_escolas
}
\keyword{datasets}

