% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_brasil_ensino_medio}
\alias{ideb_brasil_ensino_medio}
\title{IDEB do Ensino Medio Brasil}
\description{
O ideb_brasil_ensino_medio contem como atributo os valores dos indicadores utilizados para o calculo do IDEB do ensino medio para o Brasil, como as taxas de aprovacoes (por serie) e as notas da prova SAEB. Conta tambem com o valor do IDEB e as metas para os anos de 2005, 2007, 2009, 2013, 2015, 2017 e 2019. Tambem estratifica essas informacoes pela rede de ensino.
}
\usage{
ideb::ideb_brasil_ensino_medio
}
\format{
[,1] pais

[,2] rede
[,3] taxa_de_aprovacao_total 
[,4] taxa_de_aprovacao_1_ano
[,5] taxa_de_aprovacao_2_ano
[,6] taxa_de_aprovacao_3_ano
[,7] taxa_de_aprovacao_4_ano
[,8] taxa_de_aprovacao_indicador_de_rendimento
[,9] nota_SAEB_matematica
[,10] nota_SAEB_lingua_portuguesa
[,11] nota_SAEB_media_padronizada
[,12] IDEB
[,13] meta
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ideb_brasil_ensino_medio
}
\keyword{datasets}

