% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_brasil_fundamental_finais}
\alias{ideb_brasil_fundamental_finais}
\title{ideb_brasil_fundamental_finais}
\format{ideb_brasil_fundamental_finais}
\source{
ideb_brasil_fundamental_finais
}
\usage{
ideb_brasil_fundamental_finais
}
\description{
ideb_brasil_fundamental_finais
}
\keyword{datasets}

