% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_ensino_medio_brasil}
\alias{ideb_ensino_medio_brasil}
\title{IDEB do Ensino Medio Brasil}
\description{
O ideb_ensino_medio_brasil contem como atributo os valores dos indicadores utilizados para o calculo do IDEB do ensino medio para o Brasil, como as taxas de aprovacoes (por serie) e as notas da prova SAEB. Conta tambem com o valor do IDEB e as metas para os anos de 2017 e 2019. Tambem estratifica essas informacoes pela rede de ensino.
}
\usage{
ideb::ideb_ensino_medio_brasil
}
\format{
[,1] pais \cr
[,2] rede \cr
[,3] ano \cr
[,4] aprovacao_total \cr
[,5] aprovacao_1_ano \cr
[,6] aprovacao_2_ano \cr
[,7] aprovacao_3_ano \cr
[,8] aprovacao_4_ano \cr
[,9] rendimento \cr
[,10] nota_matematica \cr
[,11] nota_portugues \cr
[,12] nota_media \cr
[,13] ideb \cr
[,14] meta
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ideb_ensino_medio_brasil
}
\keyword{datasets}

