% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_ensino_medio_escolas}
\alias{ideb_ensino_medio_escolas}
\title{IDEB do Ensino Medio por Escola}
\description{
O ideb_ensino_medio_escolas contem como atributo os valores dos indicadores utilizados para o calculo do IDEB do ensino medio para as escolas do Brasil, como as taxas de aprovacoes (por serie) e as notas da prova SAEB. Conta tambem com o valor do IDEB e as metas para os anos de 2017 e 2019. Tambem estratifica essas informacoes pela rede de ensino.
}
\usage{
ideb::ideb_ensino_medio_escolas
}
\format{
[,1] sigla_uf \cr
[,2] cod_municipio \cr
[,3] nome_municipio \cr
[,4] id_escola \cr
[,5] nome_escola \cr
[,6] rede \cr
[,7] ano \cr
[,8] aprovacao_total \cr
[,9] aprovacao_1_ano \cr
[,10] aprovacao_2_ano \cr
[,11] aprovacao_3_ano \cr
[,12] aprovacao_4_ano \cr
[,13] rendimento \cr
[,14] nota_matematica \cr
[,15] nota_portugues \cr
[,16] nota_media \cr
[,17] ideb \cr
[,18] meta
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ideb_ensino_medio_escolas
}
\keyword{datasets}

