% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_escolas_ensino_medio}
\alias{ideb_escolas_ensino_medio}
\title{IDEB do Ensino Medio por Escola}
\description{
O ideb_escolas_ensino_medio contem como atributo os valores dos indicadores utilizados para o calculo do IDEB do ensino medio para as escolas do Brasil, como as taxas de aprovacoes (por serie) e as notas da prova SAEB. Conta tambem com o valor do IDEB e as metas para os anos de 2017 e 2019. Tambem estratifica essas informacoes pela rede de ensino.
}
\usage{
ideb::ideb_escolas_ensino_medio
}
\format{
[,1] pais \cr
[,2] rede \cr
[,3] taxa_de_aprovacao_total \cr
[,4] taxa_de_aprovacao_1_ano \cr
[,5] taxa_de_aprovacao_2_ano \cr
[,6] taxa_de_aprovacao_3_ano \cr
[,7] taxa_de_aprovacao_4_ano \cr
[,8] taxa_de_aprovacao_indicador_de_rendimento \cr
[,9] nota_SAEB_matematica \cr
[,10] nota_SAEB_lingua_portuguesa \cr
[,11] nota_SAEB_media_padronizada \cr
[,12] IDEB \cr
[,13] meta
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ideb_escolas_ensino_medio
}
\keyword{datasets}

