% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_escolas_fundamental_finais}
\alias{ideb_escolas_fundamental_finais}
\title{IDEB das Series Finais do Ensino Fundamental por Escolas do Brasil}
\description{
O ideb_escolas_fundamental_finais contem como atributo os valores dos indicadores utilizados para o calculo do IDEB das series finais do ensino fundamental para as escolas do Brasil, como as taxas de aprovacoes (por serie) e as notas da prova SAEB. Conta tambem com o valor do IDEB e as metas para os anos de 2005, 2007, 2009, 2011, 2013, 2015, 2017 e 2019. Tambem estratifica essas informacoes pela rede de ensino.
}
\usage{
ideb::ideb_escolas_fundamental_finais
}
\format{
[,1] pais \cr
[,2] rede \cr
[,3] aprovacao_total \cr
[,4] aprovacao_6_ano \cr
[,5] aprovacao_7_ano \cr
[,6] aprovacao_8_ano \cr
[,7] aprovacao_9_ano \cr
[,8] rendimento \cr
[,9] nota_matematica \cr
[,10] nota_portugues \cr
[,11] nota_media \cr
[,12] ideb \cr
[,13] meta
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ideb_escolas_fundamental_finais
}
\keyword{datasets}

