% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_fundamental_iniciais_brasil}
\alias{ideb_fundamental_iniciais_brasil}
\title{IDEB das series iniciais do ensino fundamental do Brasil}
\description{
O ideb_fundamental_iniciais_brasil contem como atributo os valores dos indicadores utilizados para o calculo do IDEB das series iniciais do ensino fundamental para o Brasil, como as taxas de aprovacoes (por serie) e as notas da prova SAEB. Conta tambem com o valor do IDEB e as metas para os anos de 2005, 2007, 2009, 2013, 2015, 2017 e 2019. Tambem estratifica essas informacoes pela rede de ensino.
}
\usage{
ideb_fundamental_iniciais_brasil
}
\format{
[,1] pais \cr
[,2] rede \cr
[,3] ano \cr
[,4] aprovacao_total \cr
[,5] aprovacao_1_ano \cr
[,6] aprovacao_2_ano \cr
[,7] aprovacao_3_ano \cr
[,8] aprovacao_4_ano \cr
[,9] aprovacao_5_ano \cr
[,10] rendimento \cr
[,11] nota_matematica \cr
[,12] nota_portugues \cr
[,13] nota_media \cr
[,14] ideb \cr
[,15] meta
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ideb_fundamental_iniciais_brasil
}
\keyword{datasets}

