% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideb_brasil_ensino_medio}
\alias{ideb_brasil_ensino_medio}
\title{ideb_brasil_ensino_medio}
\format{ideb_brasil_ensino_medio}
\source{
ideb_brasil_ensino_medio
}
\usage{
ideb_brasil_ensino_medio
}
\description{
ideb_brasil_ensino_medio
}
\keyword{datasets}

